Move arrcell to src/cell.
Mimic ArrowT signature for CellT
